Creating some more shaders to test efficiency of shaders
The desired efficiency is definitly not present using the current
algorithm. The fault seems to be the amount of times the engine loops
over the screen and the amount of times a new Object is created inside
the Shader.

For comparison purposes :
For now the average fps using a frame of 300 * 300 and 2 active shaders
is around 50 fps which still leaves some room to improvement.